Basic refactor + Promise chain on new http service
